+++ src/layouts/MainLayout.tsx
@@ -1,10 +1,10 @@
-import Footer from "@/components/ui/Footer";
-import { Header } from "@/components/ui/Header";
+import Footer from "@/components/shared/Footer";
+import { Header } from "@/components/shared/Header";
 import { Outlet } from "react-router-dom";
 
 const MainLayout = () => {
   return (
-    <div>
+    <div className="dark:text-white bg-white dark:bg-brand-primary">
       <Header />
       <main className="px-6 py-4 max-w-[1400px] mx-auto">
         <Outlet />
