+++ src/pages/Home/Home.tsx
@@ -9,11 +9,30 @@ import SearchInput from "@/components/ui/SearchInput";
 import Input from "@/components/ui/input";
 import Avatar from "@/components/ui/Avatar";
 import CommentItem from "@/components/ui/Comment";
+// import { useThemeStore } from "@/store/themeStore";
+// import { useEffect } from "react";
 
 const Home = () => {
+  // const theme = useThemeStore(s => s.theme);
+  // const setTheme = useThemeStore(s => s.setTheme);
+
+  // const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+  //   const checked = e.target.checked;
+  //   setTheme(checked ? "dark" : "light");
+  // };
+
+  // // DaisyUI는 input의 checked 상태를 기반으로 swap-on/off 처리하므로, 상태를 반영
+  // useEffect(() => {
+  //   const el = document.querySelector<HTMLInputElement>("input.theme-controller");
+  //   if (el) {
+  //     el.checked = theme === "dark";
+  //   }
+  // }, [theme]);
   return (
-    <div className="flex justify-center flex-col  items-center gap-4">
+    <div className="flex justify-center flex-col mt-40  items-center gap-4">
       <h1>홈 페이지</h1>
+
+      <hr />
       <h2>버튼들</h2>
       <Button intent="primary" size="lg">
         btn
